### 2019-08-01 #1616 @whedon
[PRE REVIEW]: Spruceup: fast and flexible identification, visualization, and removal of outliers from large multiple sequence alignments
**Submitting author:** @marekborowiec (<a href="http://orcid.org/0000-0001-7510-4064">Marek Borowiec</a>)
**Repository:** <a href="https://github.com/marekborowiec/spruceup" target ="_blank">https://github.com/marekborowiec/spruceup</a>
**Version:** v2019.1.1
