### 2019-08-01 #1616 @labarba
@marekborowiec —  I am the Associate Editor-in-Chief on rotation this week, and will be working to identify an editor to handle your submission.  But first, you need to add a `paper.bib` file with the references cited in your paper, in BibTeX format.
