### 2019-08-01 #1616 @labarba
Please merge this fix
https://github.com/marekborowiec/spruceup/pull/1
