### 2019-08-02 56d4236
attempted fixing figure caption rendering
