+++ paper.md
@@ -66,10 +66,4 @@ I would like to thank Philip S. Ward (UC Davis) for testing and valuable feedbac
 
 .center[
 
-![figure1](spruceup-fig1.png)
-
-.caption[
-**Figure 1:** Outlier removal from multiple sequence alignment by ``spruceup``. Divergent (complemented) sequence fragments were introduced into a simulated nucleotide data matrix (left panels), then trimmed using the ``spruceup`` workflow (right panels). Alignments visualized with AliView: https://ormbunkar.se/aliview/.
-]
-
-]
+![Outlier removal from multiple sequence alignment by ``spruceup``. Divergent (complemented) sequence fragments were introduced into a simulated nucleotide data matrix (left panels), then trimmed using the ``spruceup`` workflow (right panels). Alignments visualized with AliView: https://ormbunkar.se/aliview/.](spruceup-fig1.png)
