### 2019-08-02 #1616 @marekborowiec
Seems like I didn't set the figure caption correctly. Please try again.
