### 2019-08-02 7dc3c8a
attempted fixing figure caption rendering 2
