### 2019-08-03 #1616 @labarba
👋 @csoneson — the submitting author suggested you as handling editor. Could you take this one?

**Spruceup: fast and flexible identification, visualization, and removal of outliers from large multiple sequence alignments**
https://github.com/marekborowiec/spruceup
[ Article proof :page_facing_up: ](https://github.com/openjournals/joss-papers/blob/joss.01616/joss.01616/10.21105.joss.01616.pdf)
