### 2019-08-03 #1616 @csoneson
Ok, not exactly sure what just happened with the reference check, that looks a bit excessive...

Anyway - @marekborowiec, do you have suggestions for reviewers that would be suitable for your submission (e.g. from the list linked in the first post in this issue)?

While we are looking for reviewers, could you also fix a couple of small things in the paper:
- [ ] remove the "editor" line in the .bib entries
- [ ] be consistent in using either abbreviated or non-abbreviated journal names in the bibliography
