### 2019-08-06 7e95a32
removed editor lines from journal article entries; consistent unabbreviated journal names
