+++ paper.bib
@@ -331,7 +331,6 @@
   pages		= {e0129183},
   author	= {Md Shamsuzzoha Bayzid and Siavash Mirarab and Bastien
 		  Boussau and Tandy Warnow},
-  editor	= {Nico Cellinese},
   title		= {Weighted Statistical Binning: enabling Statistically
 		  Consistent Genome-Scale Phylogenetic Analyses},
   journal	= {{PLoS} {ONE}}
@@ -409,7 +408,6 @@
   number	= {6},
   pages		= {563--571},
   author	= {Jonathan Belmaker and Walter Jetz},
-  editor	= {Hector Arita},
   title		= {Relative roles of ecological and energetic constraints,
 		  diversification rates and region history on global species
 		  richness gradients},
@@ -561,7 +559,6 @@
   pages		= {e0161531},
   author	= {Bonnie B. Blaimer and Michael W. Lloyd and Wilson X.
 		  Guillory and Se{\'{a}}n G. Brady},
-  editor	= {Patrick O{\textquotesingle}Grady},
   title		= {Sequence Capture and Phylogenetic Utility of Genomic
 		  Ultraconserved Elements Obtained from Pinned Insect
 		  Specimens},
@@ -871,7 +868,6 @@
   month		= {jan},
   publisher	= {Oxford University Press ({OUP})},
   author	= {Marek L Borowiec},
-  editor	= {Brian Wiegmann},
   title		= {Convergent Evolution of the Army Ant Syndrome and
 		  Congruence in Big-Data Phylogenetics},
   journal	= {Systematic Biology}
@@ -898,7 +894,6 @@
   author	= {Remco Bouckaert and Joseph Heled and Denise K{\"{u}}hnert
 		  and Tim Vaughan and Chieh-Hsi Wu and Dong Xie and Marc A.
 		  Suchard and Andrew Rambaut and Alexei J. Drummond},
-  editor	= {Andreas Prlic},
   title		= {{BEAST} 2: a Software Platform for {Bayesian} Evolutionary
 		  Analysis},
   journal	= {{PLoS} Computational Biology}
@@ -1051,7 +1046,6 @@
   publisher	= {Wiley-Blackwell},
   author	= {Michael G. Branstetter and John T. Longino and Philip S.
 		  Ward and Brant C. Faircloth},
-  editor	= {Samantha Price},
   title		= {Enriching the ant tree of life: enhanced {UCE} bait set
 		  for genome-scale phylogenetics of ants and other
 		  {Hymenoptera}},
@@ -1101,7 +1095,6 @@
   number	= {6},
   pages		= {1036--1054},
   author	= {Leonidas Brikiatis},
-  editor	= {Jason Ali},
   title		= {The {De Geer}, {Thulean} and {Beringia} routes: key
 		  concepts for understanding early {Cenozoic} biogeography},
   journal	= {Journal of Biogeography}
@@ -1520,7 +1513,6 @@
   author	= {David Coty and C{\'{e}}dric Aria and Romain Garrouste and
 		  Patricia Wils and Fr{\'{e}}d{\'{e}}ric Legendre and
 		  Andr{\'{e}} Nel},
-  editor	= {Judith Korb},
   title		= {The First Ant-Termite Syninclusion in Amber with {CT}-Scan
 		  Analysis of Taphonomy},
   journal	= {{PLoS} {ONE}}
@@ -2035,7 +2027,6 @@
   pages		= {594--615},
   author	= {Mario Dos Reis and Gregg F Gunnell and Jose Barba-Montoya
 		  and Alex Wilkins and Ziheng Yang and Anne D Yoder},
-  editor	= {Simon Ho},
[…]
