### 2019-08-06 c8b94b3
moved figure up and added references section
