### 2019-08-06 #1616 @marekborowiec
I believe the following people should be able to review this:

@iimong, @gvegayon, @gavinmdouglas, @shaunpwilkinson, @giraola, @dsurujon, @Nazeeefa, @camillescott

I have also removed the editor line from bibliography entries for journal articles and attempted to make consistent use of unabbreviated journal names.

I also introduced a couple adjustments to the `paper.md` file so that it (hopefully) correctly renders the figure and references section title correctly.
