### 2019-08-06 #1616 @arfon
> I have also removed the editor line from bibliography entries for journal articles and attempted to make consistent use of unabbreviated journal names.

@marekborowiec - it would be good if you could also limit the entries in your bibtex file to just those references that you cite in this paper.
