### 2019-08-07 298958f
restricted refs to text citations
