+++ paper.bib
@@ -1,0 +1,317 @@
+@Article{	  borowiec_2015ctenophora,
+  doi		= {10.1186/s12864-015-2146-4},
+  url		= {https://doi.org/10.1186%2Fs12864-015-2146-4},
+  year		= 2015,
+  month		= {nov},
+  publisher	= {Springer Nature},
+  volume	= {16},
+  number	= {1},
+  author	= {Marek L. Borowiec and Ernest K. Lee and Joanna C. Chiu and
+		  David C. Plachetzki},
+  title		= {Extracting phylogenetic signal and accounting for bias in
+		  whole-genome data sets supports the {Ctenophora} as sister
+		  to remaining {Metazoa}},
+  journal	= {{BMC} Genomics}
+}
+
+@Article{	  brown_2017bf,
+  doi		= {10.1093/sysbio/syw101},
+  url		= {https://doi.org/10.1093%2Fsysbio%2Fsyw101},
+  year		= 2017,
+  publisher	= {Oxford University Press ({OUP})},
+  volume	= {66},
+  number	= {4},
+  pages		= {517--530},
+  author	= {Jeremy M. Brown and Robert C. Thomson},
+  title		= {Bayes factors unmask highly variable information content,
+		  bias, and extreme influence in phylogenomic analyses},
+  journal	= {Systematic Biology}
+}
+
+@Article{	  capella_gutierrez_2009trimal,
+  doi		= {10.1093/bioinformatics/btp348},
+  url		= {https://doi.org/10.1093%2Fbioinformatics%2Fbtp348},
+  year		= 2009,
+  month		= {jun},
+  publisher	= {Oxford University Press ({OUP})},
+  volume	= {25},
+  number	= {15},
+  pages		= {1972--1973},
+  author	= {S. Capella-Gutierrez and J. M. Silla-Martinez and T.
+		  Gabaldon},
+  title		= {{trimAl}: a tool for automated alignment trimming in
+		  large-scale phylogenetic analyses},
+  journal	= {Bioinformatics}
+}
+
+@Article{	  criscuolo_2010bmge,
+  doi		= {10.1186/1471-2148-10-210},
+  url		= {https://doi.org/10.1186%2F1471-2148-10-210},
+  year		= 2010,
+  publisher	= {Springer Nature},
+  volume	= {10},
+  number	= {1},
+  pages		= {210},
+  author	= {Alexis Criscuolo and Simonetta Gribaldo},
+  title		= {{BMGE} (Block Mapping and Gathering with Entropy): a new
+		  software for selection of phylogenetic informative regions
+		  from multiple sequence alignments},
+  journal	= {{BMC} Evolutionary Biology}
+}
+
+@Article{	  di_franco_2019hmmcleaner,
+  doi		= {10.1186/s12862-019-1350-2},
+  url		= {https://doi.org/10.1186%2Fs12862-019-1350-2},
+  year		= 2019,
+  month		= {jan},
+  publisher	= {Springer Nature},
+  volume	= {19},
+  number	= {1},
+  pages		= {21},
+  author	= {Di Franco, Arnaud and Poujol, Raphaël and Baurain, Denis
+		  and Philippe, Herv{\'{e}}},
+  title		= {Evaluating the usefulness of alignment filtering methods
+		  to reduce the impact of errors on evolutionary inferences},
+  journal	= {{BMC} Evolutionary Biology}
+}
+
+@Article{	  dress_2008noisy,
+  doi		= {10.1186/1748-7188-3-7},
+  url		= {https://doi.org/10.1186%2F1748-7188-3-7},
+  year		= 2008,
+  publisher	= {Springer Nature},
+  volume	= {3},
+  number	= {1},
+  pages		= {7},
[…]
