### 2019-08-07 #1616 @csoneson
👋 @iimog, @gavinmdouglas, @shaunpwilkinson, @giraola, @dsurujon, @Nazeeefa, @camillescott - would (two of) you be willing to review this submission for JOSS?
