### 2019-08-07 #1616 @marekborowiec
Is there a way to make the figure go before References?
