### 2019-08-08 #1616 @csoneson
>Is there a way to make the figure go before References?

Hm, making figures appear in the desired place in the output is certainly not always easy. From https://github.com/jgm/pandoc/issues/845 I think you can override the default placement of the figure, and it is also possible to explicitly add `\pagebreak` to force a new page. Whether that will look better in the end remains to be seen. Btw, I'm not sure you need a separate section for the figure - could it just go in the paragraph where it is mentioned?
