### 2019-08-08 #1616 @arfon
I would suggested waiting on fixing this issue until the submission is reviewed and accepted. Layouts are notoriously hard with Markdown/Pandoc and as the paper contents may change during review, we should just fix this once at the end.

Does that sounds reasonable to everyone?
