### 2019-08-08 #1616 @marekborowiec
> I would suggested waiting on fixing this issue until the submission is reviewed and accepted.

Sounds good to me. I was just wondering if my markdown file was properly formatted.
