+++ paper.bib
@@ -118,8 +118,8 @@
   volume	= {61},
   number	= {5},
   pages		= {717--726},
-  author	= {B. C. Faircloth and J. E. McCormack and N. G. Crawford and
-		  M. G. Harvey and R. T. Brumfield and T. C. Glenn},
+  author	= {Brant C. Faircloth and John E. McCormack and Nicholas G. Crawford and
+		  Michael G. Harvey and Robb T. Brumfield and Travis C. Glenn},
   title		= {Ultraconserved Elements Anchor Thousands of Genetic
 		  Markers Spanning Multiple Evolutionary Timescales},
   journal	= {Systematic Biology}
