### 2019-08-09 #1616 @gavinmdouglas
I'm familiar and interested in the topic although not an expert regarding alignment trimming - if that's ok then I'm happy to be a reviewer.
