### 2019-08-12 marekborowiec/spruceup#2 @gavinmdouglas
pip install error
Hi there,

I've reviewing spruceup for JOSS (see https://github.com/openjournals/joss-reviews/issues/1635) and I'm running into some errors when using the recommended installation commands. This is with conda 4.7.11, pip 19.1.1, and Python 3.7.4 on an Ubuntu 16.04.5 LTS system.

There are many errors, but the final lines are:

```
Installing collected packages: numpy, six, cycler, kiwisolver, pyparsing, python-dateutil, matplotlib, tqdm, psutil, scipy, treeswift, spruceup
  Running setup.py install for numpy ... done
  Running setup.py install for kiwisolver ... done
  Running setup.py install for matplotlib ... done
  Running setup.py install for psutil ... done
  Running setup.py install for scipy ... error
    ERROR: Complete output from command /home/gavin/local/miniconda3/envs/spruceup/bin/python -u -c 'import setuptools, tokenize;__file__='"'"'/tmp/pip-install-jxxayd06/scipy/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record /tmp/pip-record-806elgoe/install-record.txt --single-version-externally-managed --compile:
    ERROR:
    Note: if you need reliable uninstall behavior, then install
    with pip instead of using `setup.py install`:

      - `pip install .`       (from a git repo or downloaded source
                               release)
      - `pip install scipy`   (last SciPy release on PyPI)

    Traceback (most recent call last):
      File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/core/__init__.py", line 16, in <module>
        from . import multiarray
    ImportError: cannot import name 'multiarray' from 'numpy.core' (/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/core/__init__.py)

    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/tmp/pip-install-jxxayd06/scipy/setup.py", line 492, in <module>
        setup_package()
      File "/tmp/pip-install-jxxayd06/scipy/setup.py", line 468, in setup_package
        from numpy.distutils.core import setup
      File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/__init__.py", line 142, in <module>
        from . import add_newdocs
      File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/add_newdocs.py", line 13, in <module>
        from numpy.lib import add_newdoc
      File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/lib/__init__.py", line 8, in <module>
        from .type_check import *
      File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/lib/type_check.py", line 11, in <module>
        import numpy.core.numeric as _nx
      File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/core/__init__.py", line 26, in <module>
        raise ImportError(msg)
    ImportError:
    Importing the multiarray numpy extension module failed.  Most
    likely you are trying to import a failed build of numpy.
    If you're working with a numpy git repo, try `git clean -xdf` (removes all
    files not under version control).  Otherwise reinstall numpy.

    Original error was: cannot import name 'multiarray' from 'numpy.core' (/home/gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/numpy/core/__init__.py)

ERROR: Command "/home/gavin/local/miniconda3/envs/spruceup/bin/python -u -c 'import setuptools, tokenize;__file__='"'"'/tmp/pip-install-jxxayd06/scipy/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' install --record /tmp/pip-record-806elgoe/install-record.txt --single-version-externally-managed --compile" failed with error code 1 in /tmp/pip-install-jxxayd06/scipy/
```

See the attached file for the full error log and the commands I ran at the beginning.
[spruceup_install_error.txt](https://github.com/marekborowiec/spruceup/files/3492958/spruceup_install_error.txt)
