### 2019-08-12 marekborowiec/spruceup#2 @gavinmdouglas
Note that running `pip install` with the `--no-cache-dir` option did not resolve this issue.
