### 2019-08-12 dd7718f
changed python and dependency version requirements, bumped version to 2019.1.1b
