### 2019-08-12 4313437
changed python and dependency version requirements, bumped version to 2019.1.2
