### 2019-08-12 4c85f2c
changed Python 3.7 requirement to 3.6 and explained potential issue with 3.7.4
