### 2019-08-12 marekborowiec/spruceup#2 @marekborowiec
I am not able to replicate this exact error but I'm unable to install `spruceup` with Python 3.7.4 either, although 3.7.3 works fine. After playing around with different Python and dependency versions, I updated the distribution setup.py and documentation, allowing Python 3.6 and up but disallowing Python 3.7.4 for now. Please try again with other version and let me know how it goes.
