### 2019-08-13 marekborowiec/spruceup#2 @iimog
I can confirm that the `conda` and `pip` installation method works on Ubuntu 19.04 with Python 3.7.3
