### 2019-08-13 marekborowiec/spruceup#2 @gavinmdouglas
Thanks - the installation works on my system now with the new Python 3.6 instructions.
