### 2019-08-14 marekborowiec/spruceup#4 @iimog
Example command
As part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/1635) I ran the example command and I have some comments:
1. Installing `spruceup` through `pip` does not copy the `examples` directory to the users computer. So it would help to let the users know that in order to try the example they have to download the `tar.gz` or clone the repo (or at least download the files in `examples` from GitHub).
2. Running the example command exactly as stated in the README (from the `spruceup` directory not the `examples` directory) results in:
`Sorry, could not read input alignment file "test_aln_nt.fasta": No such file or directory`
Running the command from the `examples` directory works.
3. The example config does not work for single core machines. Running on a single core machine results in:
`You specified more (2) compute cores than are available (1). Exiting.`
I don't think this is a big problem as changing the setting to 1 is straightforward. However, 1 core is enough for this demo so you might as well change this setting to 1 by default in the demo to avoid this issue.
