### 2019-08-15 marekborowiec/spruceup#5 @gavinmdouglas
JOSS review
Related to JOSS review https://github.com/openjournals/joss-reviews/issues/1635.

### Major comment

`spruceup` seems like it will be a useful tool for the field. However, I think numerous additions to the documentation will make it a more useful resource.

### Specific comments

- When briefly describing spruceup at the top of the README it would be good to add a sentence stating what research gap this tool addresses. I.e. what are the advantages of this tool compared to the other tools you mention in your paper? This would help address this requirement: `A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?`.

- There should be an official github release matching the pip release in the future so that it's clear what exact version on GitHub matches the source code distributed with pip.

- Installation instructions should be re-ordered so that recommended conda/pip instructions are first and "conda activate spruceup" should also be listed clearly as a separate line so that it can be copied and pasted more easily.

- I recommend that the input option / argument handling is improved. For instance, argparse could be used, which would avoid the need to use a configuration file and provide help documentation to the user. If you want to stick with the configuration file approach you need to at least point the user to where they can find the documentation for using the script if they run the command with no options or if they input "--help" or "-h" (see next point).

- At the very least the "--version" and "--help" options should be added to the script.

- I received the below error when installing the tool on OS X. You should provide instructions for Mac users if you intend the tool to be used on OS X.

```
    Traceback (most recent call last):
      File "/Users/Gavin/local/miniconda3/envs/spruceup/bin/spruceup.py", line 18, in <module>
        import matplotlib.pyplot as plt
      File "/Users/Gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/matplotlib/pyplot.py", line 2374, in <module>
        switch_backend(rcParams["backend"])
      File "/Users/Gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/matplotlib/pyplot.py", line 207, in switch_backend
        backend_mod = importlib.import_module(backend_name)
      File "/Users/Gavin/local/miniconda3/envs/spruceup/lib/python3.7/importlib/__init__.py", line 127, in import_module
      File "/Users/Gavin/local/miniconda3/envs/spruceup/lib/python3.7/site-packages/matplotlib/backends/backend_macosx.py", line 14, in <module>
        from matplotlib.backends import _macosx
    ImportError: Python is not installed as a framework. The Mac OS X backend will not be able to function correctly if Python is not installed as a framework. See the Python documentation for more information on installing Python as a framework on Mac OS X. Please either reinstall Python as a framework, or try one of the other backends. If you are using (Ana)Conda please install python.app and replace the use of 'python' with 'pythonw'. See 'Working with Matplotlib on OSX' in the Matplotlib FAQ for more information.
```

- Add instructions for how to run unittests if installing from source

- Recommend changing criterion setting in example config file to mean if that is the recommended option.

- More details are needed for the example usage dataset. In particular, it's not clear whether this is a realistic dataset or not.

- The example command also failed for me with `Sorry, could not read input alignment file "test_aln_nt.fasta": No such file or directory` (it worked when I ran it in the `examples` folder)

- "?" is not an IUPAC character (https://www.bioinformatics.org/sms/iupac.html) and may cause confusion for users and downstream programs. Is it intended to represent a gap or an N character? The appropriate IUPAC character should be output instead if possible otherwise a clearer explanation needs to be given when describing the output.

### Manuscript edits

- First sentence is very long - recommend to remove "by leveraging the large amount of data available" to be simply "from multiple sequence alignments".

- Usage of "phylogenomic" might be misleading since this tool can work with any MSA right? I don't think this is a standard term so it might cause some confusion.

- By "incorrect orthology prediction" do you mean homoplasy?

- "Multiple Sequence Alignment" should be written in the same case in both the overview and summary sections.

- Add "the" before "evolutionary history of genetic material"

- Add "they are" before "derived from the same"

- Replace "Because of complexities" with "Due to the complexities"

- OTU should be defined on first usage

- I'm not familiar with the term "subtended" - just double checking that it's the intended word

- Could use a sentence stating what users you envision using this tool - do you think this would be best for newcomers or experienced phylogenetics people or both?

- Do the sequences need to be representative sequences (i.e. OTUs)? Will the tool have trouble if there are many nearly identical sequences? If not then specifying OTUs rather than just saying "sample" or "sequence" may just cause confusion.

- I think it's more accurate to say that it's written in Python 3, not 3.7.

- A lot of the rows in Figure 1 are redundant - I think removing most of them would make this a lot easier to visualize (and you can simply mention that more sequences were actually included originally). Also more details are needed such as what "?" corresponds to.
