### 2019-08-19 marekborowiec/spruceup#6 @iimog
Unit tests
As pointed out by @gavinmdouglas in #5 the unit tests should be documented. You may also consider using continuous integration such as https://travis-ci.org/
I was able to run the `simple_tests.py` from the `tests` folder (after `conda install nose`) with:
```
$ nosetests simple_tests.py
....................................
Ran 36 tests in 0.005s

OK
```
However, the `metamorphic_tests.py` fails with:
```
$ nosetests metamorphic_tests.py
FFFF
FAIL: tests.metamorphic_tests.metamorphic_tests.test_correct_removal_baseline
Traceback (most recent call last):
  File "/home/ankenbrand_m/miniconda3/envs/spruceup/lib/python3.7/site-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/home/ankenbrand_m/tmp/spruceup/tests/metamorphic_tests/metamorphic_tests.py", line 89, in test_correct_removal_baseline
    assert_true('Sites removed: 20500' in str(baseline.stderr))
AssertionError: False is not true

FAIL: tests.metamorphic_tests.metamorphic_tests.test_correct_removal_complemented
Traceback (most recent call last):
  File "/home/ankenbrand_m/miniconda3/envs/spruceup/lib/python3.7/site-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/home/ankenbrand_m/tmp/spruceup/tests/metamorphic_tests/metamorphic_tests.py", line 99, in test_correct_removal_complemented
    assert_true('Sites removed: 20500' in str(complemented.stderr))
AssertionError: False is not true

FAIL: tests.metamorphic_tests.metamorphic_tests.test_correct_removal_loci_randomized
Traceback (most recent call last):
  File "/home/ankenbrand_m/miniconda3/envs/spruceup/lib/python3.7/site-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/home/ankenbrand_m/tmp/spruceup/tests/metamorphic_tests/metamorphic_tests.py", line 109, in test_correct_removal_loci_randomized
    assert_true('Sites removed: 20500' in str(scrambled_loci.stderr))
AssertionError: False is not true

FAIL: tests.metamorphic_tests.metamorphic_tests.test_correct_removal_taxa_randomized
Traceback (most recent call last):
  File "/home/ankenbrand_m/miniconda3/envs/spruceup/lib/python3.7/site-packages/nose/case.py", line 197, in runTest
    self.test(*self.arg)
  File "/home/ankenbrand_m/tmp/spruceup/tests/metamorphic_tests/metamorphic_tests.py", line 122, in test_correct_removal_taxa_randomized
    assert_true('Sites removed: 20500' in str(scrambled_taxa.stderr))
AssertionError: False is not true

Ran 4 tests in 2.136s

FAILED (failures=4)
```
This issue is also related to openjournals/joss-reviews#1635
