### 2019-08-19 marekborowiec/spruceup#5 @iimog
In addition to the issues I opened separately (#3, #4, #6) and to the points @gavinmdouglas already raised I'd like to add:
 - I agree that this tool is likely of high utility to the community but can be further improved
 - Although there are some basic guidelines for contributing, these suggestions can make it even simpler: https://github.com/marekborowiec/spruceup/community

Manuscript:
 - I'm a bit confused by the usage of "sample", "sequence" and "OTU" to refer to the alignment rows. Maybe it would be better to stick to one of them or clearly distinguish the different meanings.
 - I don't see how "*Lack of tools for automated detection of these sequences **means that** few researchers attempt manual alignment curation [...]*". I'd rather assume that the availability of automated tools means that few use manual curation. I agree that many will not address it at all, though.
 - Duplicated `it` in "*It it is also capable [...]*"
