### 2019-09-03 #1635 @marekborowiec
Hello all,

Thank you for checking in, @csoneson; I'm working on it!
@iimog and @gavinmdouglas: thank you both for constructive reviews! Your suggestions will help me improve the tool. Please bear with me as I'm slowly getting through them; I might be asking some clarification questions along the way!
