### 2019-09-03 ebfc2d5
moved conda installation instructions to be first
