### 2019-09-04 d7be2a2
help and config file arguments now handled by argparse
