+++ __init__.py
@@ -1,0 +1,2 @@
+__version_info__ = ('2019','2','1')
+__version__ = '.'.join(__version_info__)
