+++ README.md
@@ -21,7 +21,7 @@ This script uses [numpy](http://www.numpy.org/), [scipy](https://www.scipy.org/i
 You can download a zipped GitHub repository, clone it if you have `git` installed on your system, or install using [pip](https://pip.pypa.io/en/latest/installing.html) (recommended) from the [Python Package Index](https://pypi.python.org/pypi/spruceup).
 
 
-`spruceup` requires you have Python version 3.6 or newer. You can use `spruceup` with early releases of Python 3.7, but as of today (12 August 2019), you may experience trouble installing it if using Python 3.7.4. Dependencies should be installed automatically. If your system does not have Python version 3.6 or newer you will need to [download and install it](http://www.python.org/downloads/). On Linux-like systems (including Ubuntu) you can install it from the command line using
+`spruceup` requires you have Python version 3.6 or newer. You can use `spruceup` with early releases of Python 3.7, but as of today (12 August 2019), you may experience trouble installing it if using Python 3.7.4. Dependencies should be installed automatically. If your system does not have Python version 3.6 or newer you will need to [download and install it](http://www.python.org/downloads/). On Linux-like systems (including Ubuntu) you can install it from the command line using:
 
 ```bash
 sudo apt-get install python3.6
@@ -75,7 +75,7 @@ The `analysis` category defines parameters used to analyze and clean up your ali
 
 `fraction` signifies proportion of OTUs/samples that will be used to calculate average distance in each window. With fraction set to `1.0` distances for each OTU will be calculated against all other OTUs in the alignment. With fraction set to `0.5` distances for each OTU will be calculated against a random sample representing 50% of OTUs in the alignment. Lowering this number will help to speed up calculations in alignments with large numbers of taxa.
 
-`criterion` chooses how outlier distances will be determined. If you are using `mean` (recommended), simple multiple of those values computed for each OTU will be considered cutoffs for identifying outliers. `lognorm` means that a [lognormal distribution](https://en.wikipedia.org/wiki/Log-normal_distribution) will be fitted to your distance data for each OTU and cutoffs will be determined by specifying quantile of observations above which sequence will be considered outliers.
+`criterion` chooses how outlier distances will be determined. If you are using `mean`, simple multiple of those values computed for each OTU will be considered cutoffs for identifying outliers. `lognorm` means that a [lognormal distribution](https://en.wikipedia.org/wiki/Log-normal_distribution) will be fitted to your distance data for each OTU and cutoffs will be determined by specifying quantile of observations above which sequence will be considered outliers.
 
 `cutoffs` specifies multiple values considered as cutoffs. If you are using `mean` as your criterion, use multiples of those values, for example `5,30` etc. For `lognorm` criterion use fractions of `1`, for example `0.9,0.995` etc.  Default values that should work for most alignments are `0.9,0.95,0.97,0.99,0.995,0.999` for `lognorm` and `5,10,15,18,20,25,30` for `mean`. You can always trim with additional criteria and cutoffs after the initial analysis (see point 5 below under Interpreting the output). If the alignment contains many saturated or poorly aligned sites, a low setting may result in huge amount of data being trimmed from the original alignment. This is time-consuming and you may want to trim your alignment with a more stringent 'block' method before using `spruceup` or remove lower cutoff values from the list.
 
