### 2019-09-05 fc17bb4
restructured criterion and cutoffs sections
