### 2019-09-11 7a67e87
more JOSS review fixes
