+++ paper.md
@@ -14,49 +14,49 @@ authors:
 affiliations:
  - name: Department of Entomology, Plant Pathology and Nematology, University of Idaho
    index: 1
-date: 19 July 2019
+date: 6 September 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-``spruceup`` is an open-source Python software package that utilizes a novel algorithm and allows flexible identification, visualization, and removal of abnormal sequences by leveraging the large amount of data available in phylogenomic multiple sequence alignments (MSAs). The ``spruceup`` algorithm aims to solve the problem of individual poorly aligned sequences rarely addressed by other alignment trimming tools. It relies on computing distance metrics in a sliding window along an MSA and comparing them to metrics across all windows for each sample. While the program is designed with trimming phylogenetic datasets in mind, its output can be used for any task requiring computing of distances along MSAs with custom parameters.
+``spruceup`` is an open-source Python software package that utilizes a novel algorithm and allows flexible identification, visualization, and removal of abnormal sequences from multiple sequence alignments (MSAs). The ``spruceup`` algorithm aims to solve the problem of individual poorly aligned sequences rarely addressed by other alignment trimming tools. It relies on computing distance metrics in a sliding window along an MSA and comparing them to metrics across all windows for each sample. While the program is designed with trimming phylogenetic datasets in mind, its output can be used for any task requiring computing of distances along MSAs with custom parameters.
 
-Evolutionary inference is increasingly reliant on large MSAs consisting of millions of sites from hundreds or thousands of samples. MSAs used for inference of evolutionary relationships should contain homologous nucleotides or amino acids (alignment columns) from different samples (alignment rows). However, MSAs may contain errors that violate this assumption of homology, resulting from sample cross-contamination, incorrect orthology prediction, high divergence leading to insufficient information for accurate alignment, or other failures of alignment algorithms. These errors have been well-documented and shown to negatively impact phylogenetic inference. While many tools exist for trimming of poorly-aligned MSA columns, few allow identification of individual misaligned sequences. Tools for removal of poorly aligned columns address issues impacting most samples in the MSA but are not designed to identify misalignment limited to a single or few sequences (rows). Failure to remove such individual misaligned sequences introduces errors into downstream analyses. Lack of tools for automated detection of these sequences means that few researchers attempt manual alignment curation and many simply do not address them. Although visual inspection of data is always desirable, manual error correction is unsustainable with the alignment sizes used in modern phylogenetic studies. ``spruceup`` is intended to complement commonly used tools that filter poorly-aligned MSA sites to further improve alignment quality.
[…]
