### 2019-09-16 528689c
now substituting missing data with - instead of ?
