### 2019-09-16 a920daa
now substituting missing data with - instead of ?
