+++ paper.bib
@@ -1,3 +1,17 @@
+
+@Article{blaimer_2016acropyga,
+  doi = {10.1016/j.ympev.2016.05.030},
+  url = {https://doi.org/10.1016%2Fj.ympev.2016.05.030},
+  year = 2016,
+  month = {sep},
+  publisher = {Elsevier {BV}},
+  volume = {102},
+  pages = {20--29},
+  author = {Bonnie B. Blaimer and John S. LaPolla and Michael G. Branstetter and Michael W. Lloyd and Se{\'{a}}n G. Brady},
+title = {Phylogenomics, biogeography and diversification of obligate mealybug-tending ants in the genus \textit{Acropyga}},
+  journal = {Molecular Phylogenetics and Evolution}
+}
+
 @Article{	  borowiec_2015ctenophora,
   doi		= {10.1186/s12864-015-2146-4},
   url		= {https://doi.org/10.1186%2Fs12864-015-2146-4},
