### 2019-09-16 15ec96f
expanded figure caption
