### 2019-09-18 7968e44
added instructions for installing from source, replaced references to OTU with sample and clarified meaning
