+++ README.md
@@ -18,7 +18,7 @@ This script uses [numpy](http://www.numpy.org/), [scipy](https://www.scipy.org/i
 
 ## Installation and requirements
 
-You can download a zipped GitHub repository, clone it if you have `git` installed on your system, or install using [pip](https://pip.pypa.io/en/latest/installing.html) (recommended) from the [Python Package Index](https://pypi.python.org/pypi/spruceup).
+You can download a zipped GitHub repository, clone it if you have `git` installed on your system, or install using [pip](https://pip.pypa.io/en/latest/installing.html) from the [Python Package Index](https://pypi.python.org/pypi/spruceup). Cloning the repository will also download example files used in the tutorial below. 
 
 
 `spruceup` requires you have Python version 3.6 or newer. You can use `spruceup` with early releases of Python 3.7, but as of today (12 August 2019), you may experience trouble installing it if using Python 3.7.4. Dependencies should be installed automatically. If your system does not have Python version 3.6 or newer you will need to [download and install it](http://www.python.org/downloads/). On Linux-like systems (including Ubuntu) you can install it from the command line using:
@@ -32,12 +32,22 @@ It is best to create a new environment before installing spruceup, either using 
 conda create --name spruceup python=3.6
 ```
 
-Now activate your environment (with `conda` you would type `conda activate spruceup`) and install:
+Now activate your environment (with `conda` you would type `conda activate spruceup`) and install. If you want to download example files, clone or download this repository with:
+```bash
+git clone https://github.com/marekborowiec/spruceup.git
+```
+
+You can then install `spruceup` from within this cloned repository:
+```bash
+python setup.py install
+```
+
+`spruceup` is also available on [Python Package Index](https://pypi.org/) and you can install it via `pip` without example files:
 ```bash
 pip install spruceup
 ```
 
-You can also use a different environment manager, such as `venv`:
+To set up the new environment you can also use a different environment manager, such as `venv`, instead of `conda`:
 ```bash
 python3 -m venv spruceup
 ```
@@ -60,7 +70,7 @@ The `input` category defines parameters of the input alignment and its type.
 
 `distances_object_file` is the file name of an existing distance object. Because sometimes you will want to adjust cutoffs or cutoff criterion and computing distances is the most time-consuming part of the analysis, `spruceup` saves a `json` format file with distances from each analysis. By default this is blank, but if you do have a distance file from a previous analysis and you want to trim your alignment with new cutoffs, supply the `json` file name here. `spruceup` will then run with new trimming cutoffs and/or criterion but without the need to re-calculate distances.
 
-`guide_tree` is a phylogram or cladogram `NEWICK` format file to be used as a guide tree. The tree can be inferred using any method and can be fully resolved or contain polytomies. If you do not supply a guide tree the program will still run but without phylogeny it will have less information to identify misaligned sequences. This is particularly important in vartiable algnments with distantly related samples, where it is more difficult to distinguish genuinely variable sequences from misaligned fragments. In cases where you suspect and are mainly concerned with samples with spuriously long terminal branches it is adivsable to supply topology-only (cladogram) guide tree and/or run the program without a guide tree. 
+`guide_tree` is a phylogram or cladogram `NEWICK` format file to be used as a guide tree. The tree can be inferred using any method and can be fully resolved or contain polytomies. If you do not supply a guide tree the program will still run but without phylogeny it will have less information to identify misaligned sequences. This is particularly important in vartiable algnments with distantly related samples (individual sequences representing individual, taxon, OTU etc.), where it is more difficult to distinguish genuinely variable sequences from misaligned fragments. In cases where you suspect and are mainly concerned with samples with spuriously long terminal branches it is adivsable to supply topology-only (cladogram) guide tree and/or run the program without a guide tree. 
 
 ### [analysis]
[…]
