### 2019-09-23 7411300
added init to scripts
