### 2019-09-23 a15694f
added statement about example alignment, minor adjustments
