### 2019-09-23 51e3930
minor adjustments
