### 2019-09-24 08a2240
added instructions for testing
