### 2019-09-24 8b4e569
fixed issues with metamorphic testing code
