+++ tests/metamorphic_tests/test_metamorphic.py
@@ -1,0 +1,144 @@
+#! /usr/bin/env python3
+import glob, os, subprocess
+
+from nose.tools import assert_true
+
+# the data is a simulated nucleotide alignment with 41 taxa, 100 loci each 500 nt long
+# with every taxon having 5 randomly selected loci complemented
+# making it easy to flag the outliers
+# and should result in 20,500 nt or 1% total seq removed
+
+# metamorphic tests expect that the results will be the same if
+# 1) all sequences are complemented
+# 2) order of loci in alignment is scrambled
+# 3) order of taxa in alignment is scrambled
+
+# these tests are likely to take several minutes
+
+wd = os.path.dirname(os.path.realpath(__file__)) 
+
+def setup_module():
+    subprocess.run(['./complement.py', 'simulated.fas', 'fasta'], cwd=wd)
+    subprocess.run(['cp', 'simulation.conf', 'complemented-simulation.conf'], cwd=wd)
+    subprocess.run(
+        [
+            'sed',
+            '-i',
+            's/simulated/complemented-simulated/g',
+            'complemented-simulation.conf',
+        ],
+        cwd=wd
+    )
+    subprocess.run(
+        ['./scramble-loci.py', 'complemented-simulated.fas', 'fasta', '500'],
+        cwd=wd
+    )
+    subprocess.run(
+        ['cp', 'simulation.conf', 'scrambled-loci-complemented-simulation.conf'],
+        cwd=wd
+    )
+    subprocess.run(
+        [
+            'sed',
+            '-i',
+            's/simulated/scrambled-loci-complemented-simulated/g',
+            'scrambled-loci-complemented-simulation.conf',
+        ], 
+        cwd=wd
+    )
+    subprocess.run(
+        [
+            './scramble-taxa.py',
+            'scrambled-loci-complemented-simulated.fas',
+            'fasta',
+        ],
+        cwd=wd
+    )
+    subprocess.run(
+        [
+            'cp',
+            'simulation.conf',
+            'scrambled-taxa-scrambled-loci-complemented-simulation.conf',
+        ],
+        cwd=wd
+    )
+    subprocess.run(
+        [
+            'sed',
+            '-i',
+            's/simulated/scrambled-taxa-scrambled-loci-complemented-simulated/g',
+            'scrambled-taxa-scrambled-loci-complemented-simulation.conf',
+        ],
+        cwd=wd
+    )
+
+
+def teardown_module():
+    print('Cleaning up directory ...')
+    pngs = glob.glob('{}/*.png'.format(wd))
+    cnfs = glob.glob('{}/*-simulation.conf'.format(wd))
+    outs = glob.glob('{}/*lognorms*'.format(wd))
+    logs = glob.glob('{}/*.log'.format(wd))
+    jsns = glob.glob('{}/*.json'.format(wd))
+    mtcs = glob.glob('{}/*-simulated.fas'.format(wd))
+    remove_wrapper(pngs)
+    remove_wrapper(cnfs)
+    remove_wrapper(outs)
+    remove_wrapper(logs)
+    remove_wrapper(jsns)
+    remove_wrapper(mtcs)
+    
+
+def test_correct_removal_baseline():
+    baseline = subprocess.run(
+        ['spruceup.py', 'simulation.conf'],
+        cwd=wd,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+    )
+    assert_true('Sites removed: 20500' in str(baseline.stderr))
+    assert_true('Removed 1.00%' in str(baseline.stderr))
+
+
+def test_correct_removal_complemented():
+    complemented = subprocess.run(
[…]
