### 2019-09-26 #1635 @marekborowiec
Thanks for checking in @csoneson. Still working on it. I'm almost done with the long laundry list of problems in Issue #5. My schedule becomes better at the end of this week and I hope I'll be able to address all the remaining issues next week.
