### 2019-10-01 9b4efa1
removed nosetests from setuptools integration
