+++ setup.py
@@ -7,6 +7,7 @@ here = path.abspath(path.dirname(__file__))
 
 with open(path.join(here, 'README.md')) as f:
     long_description = f.read()
+
 setup(
     name='spruceup',
     version=__version__,
@@ -15,11 +16,9 @@ setup(
     description='A module for lexible identification, visualization, and removal of outliers from large multiple sequence alignments',
     long_description=long_description,
     long_description_content_type='text/markdown',
-    url='https://github.com/marekborowiec/seq-spruceup',
+    url='https://github.com/marekborowiec/spruceup',
     keywords='alignment trimming outliers phylogenetics phylogenomics genomics bioinformatics',
     packages=['spruceup'],
-    package_dir={'spruceup':
-                 'spruceup'},
     include_package_data=True,
     scripts=['./spruceup/spruceup.py', './spruceup/aln_parsing.py', './spruceup/aln_writing.py', './__init__.py'],
     python_requires='>=3.6.0, !=3.7.4',
@@ -40,6 +39,4 @@ setup(
         'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
         'Operating System :: OS Independent',
     ],
-    test_suite='nose.collector',
-    tests_require='nose==1.3.7',
 )
