### 2019-10-01 ce40199
added testing instructions
