### 2019-10-01 339b3ba
added statement about using concatenated alignments vs single locus
