### 2019-10-03 fbd6948
tests now with unittest and integrated with setuptools
