+++ tests/metamorphic_tests/complement.py
@@ -1,15 +1,15 @@
 #! /usr/bin/env python3
 
 # coding: utf-8
+import os
 from sys import argv 
 
-from spruceup import aln_parsing, aln_writing
-
+from spruceup import aln_writing
 # load in alignments
 # for each taxon reverse complement 1 seq in all 5 different loci
 # write modified alignments 
 
-
+wd = os.path.dirname(os.path.realpath(__file__)) 
 # reverse complement
 def complement(dna_string):
     replacement1 = dna_string.replace('A', 't')
@@ -27,17 +27,9 @@ def complement_alignment(aln_dict):
 
 def complement_wrapper(aln_tuple):
     aln_name, aln_dict = aln_tuple
+    f_name = os.path.basename(aln_name)
+    new_aln_name = f'{wd}/complemented-{f_name}'
     compl_aln = complement_alignment(aln_dict)
     aln_writing.write_alignment_file(
-        compl_aln, 'fasta', f'complemented-{aln_name}', 'dna'
+        compl_aln, 'fasta', new_aln_name, 'dna'
         )
-
-
-def get_parsed_aln(aln_filename, file_format):
-    aln_tuple = aln_parsing.parse_alignment(aln_filename, file_format)
-    return aln_tuple
-
-
-script, aln_filename, file_format = argv
-aln_tuple = get_parsed_aln(aln_filename, file_format)
-complement_wrapper(aln_tuple)
