### 2019-10-03 6830e30
updated for use as module, new testing instructions
