+++ README.md
@@ -8,7 +8,7 @@
 
 Tools to discover, visualize, and remove outlier sequences in large multiple sequence alignments.
 
-`Spruceup` is a Python tool for biologists (bioinformaticians, phylogeneticists, evolutionary biologists) doing inference on phylogenomic sequence alignments. It allows discovery and removal of individual poorly aligned sequences or sequence fragments (alignment rows), which is different from the problem of poorly aligned sequence blocks (alignment columns) commonly addressed by alignment trimming software.
+`spruceup` is a Python tool for biologists (bioinformaticians, phylogeneticists, evolutionary biologists) doing inference on phylogenomic sequence alignments. It allows discovery and removal of individual poorly aligned sequences or sequence fragments (alignment rows), which is different from the problem of poorly aligned sequence blocks (alignment columns) commonly addressed by alignment trimming software.
 
 If you are using this program, please cite [this publication](link):
 ```
@@ -27,12 +27,13 @@ You can download a zipped GitHub repository, clone it if you have `git` installe
 sudo apt-get install python3.6
 ```
 
-It is best to create a new environment before installing spruceup, either using `conda`:
+It is best to create a new environment before installing spruceup. If you are using the [conda package manager](https://www.anaconda.com/distribution/), type:
 ```bash
 conda create --name spruceup python=3.6
 ```
+And follow instuctions on the prompt.
 
-Now activate your environment (with `conda` you would type `conda activate spruceup`) and install. If you want to download example files, clone or download this repository with:
+Now activate your environment (with `conda` you would type `conda activate spruceup`) and install. If you want to download example files, simply download and unzip this repository or, if you have git, clone it with:
 ```bash
 git clone https://github.com/marekborowiec/spruceup.git
 ```
@@ -47,7 +48,7 @@ python setup.py install
 pip install spruceup
 ```
 
-To set up the new environment you can also use a different environment manager, such as `venv`, instead of `conda`:
+If you do not wish to install `conda`, you can set up a new environment using a different package manager, such as `venv`:
 ```bash
 python3 -m venv spruceup
 ```
@@ -55,22 +56,22 @@ python3 -m venv spruceup
 ## Interface
 Once you successfully installed `spruceup` you will need 1) an alignment in `FASTA`, `PHYLIP` or `NEXUS` format, 2) (optional) a guide tree for your alignment in `NEWICK` format, and 3) configuration file to run the program. To run the program from the command line you can type:
 ```bash
-spruceup.py my-configuration-file.conf
+python -m spruceup my-configuration-file.conf
 ```
-Directory `examples` contains a template configuration file. It has the following fields:
+Directory `examples` contains a template configuration file (`/examples/config_example.conf`). This file should be a template for your own analyses. It has the following fields:
 
 ### [input]
 The `input` category defines parameters of the input alignment and its type.
 
-`input_file_name` is the file path of the alignment to be processed. *This should be a relatively large, concatenated alignment*. The `spruceup` algorithm works the better the more data it has, so if you have a phylogenomic dataset consisting of many single-locus alignments, you should first concatenate them. You can then split the processed/trimmed alignment back into single-locus alignments using another utility such as [AMAS](https://doi.org/10.7717/peerj.1660).
+`input_file_name` is the file path of the alignment to be processed. This can be provided as a file name, if the config is in the same directory as alignment and `spruceup` command will be ran from there as well. Alternatively, you can provide the absolute path. *__This should be a relatively large, concatenated alignment__*. The `spruceup` algorithm works the better the more data it has, so if you have a phylogenomic dataset consisting of many single-locus alignments, you should first concatenate them. You can then split the processed/trimmed alignment back into single-locus alignments using another utility such as [AMAS](https://doi.org/10.7717/peerj.1660).
 
[…]
