+++ spruceup/__main__.py
@@ -1,0 +1,184 @@
+import argparse
+import logging
+import os
+import time
+
+import psutil
+
+from spruceup import aln_parsing, aln_writing, spruceup
+from spruceup.__init__ import __version__
+
+parser = argparse.ArgumentParser(usage='spruceup.py <config_file_name>')
+
+parser.add_argument(
+    'config_file_name',
+    type = str,
+    help = 'Configuration file with spruceup run parameters.'
+    )
+parser.add_argument(
+    '-v',
+    '--version', 
+    action='version',
+    version='%(prog)s {version}'.format(version=__version__),
+    help='show the version number and exit'
+    )
+
+
+def print_mem():
+    """Print memory usage using psutil module."""
+    process = psutil.Process(os.getpid())
+    mem = process.memory_info().rss / 1_000_000
+    logging.info('Used {0:.2f} MB memory\n'.format(mem))
+
+
+def analyze(
+    alignment_file_name,
+    tree_file_name,
+    input_file_format,
+    data_type,
+    window_size,
+    overlap,
+    cores,
+    method,
+    fraction,
+):
+    """Load, parse, and analyze the data."""
+    logging.info('Parsing alignment {} ...\n'.format(alignment_file_name))
+    aln_tuple = aln_parsing.parse_alignment(
+        alignment_file_name, input_file_format
+    )
+    aln_name, aln_dict = aln_tuple
+    no_missing_ambiguous_dict = spruceup.replace_missing_in_dict(aln_dict, data_type)
+    if tree_file_name == None:
+        logging.info('Did not find guide tree, continuing ...\n')
+        tree_dists = None
+    else:
+        logging.info('Reading in guide tree {} ...\n'.format(tree_file_name))
+        tree_dists = spruceup.get_tree_dist_dict(tree_file_name)
+    windows = spruceup.get_windows(no_missing_ambiguous_dict, window_size, overlap)
+    all_distances = spruceup.distances_wrapper(
+        windows, tree_dists, cores, data_type, method=method, fraction=fraction
+    )
+    taxa_distances = spruceup.dist_taxa_wrapper(all_distances)
+    mean_aln_distances = spruceup.mean_distances_wrapper(taxa_distances)
+    mean_taxon_distances = spruceup.dists_per_taxon(mean_aln_distances)
+    spruceup.write_distances_dict(mean_taxon_distances, method, window_size, overlap)
+    return (aln_tuple, mean_taxon_distances)
+
+
+def output_loop(
+    untrimmed_alignment,
+    distances,
+    window_size,
+    method,
+    criterion,
+    cutoffs,
+    manual_cutoffs,
+    report_file_name,
+    out_format,
+    out_file_name,
+    data_type,
+):
+    """Write and plot output files from analysis."""
+    alignment_sites = spruceup.get_alignment_size(untrimmed_alignment)
+    cutoff_floats = [float(cutoff_string) for cutoff_string in cutoffs]
+    for cutoff in cutoff_floats:
+        logging.info(
+            'Finding outliers for {} {}s cutoff ...'.format(cutoff, criterion)
+        )
+        outliers = spruceup.get_outliers_wrapper(
+            distances, window_size, method, criterion, cutoff, manual_cutoffs
+        )
+        sites_removed, trimmed_alignment = spruceup.remove_outliers(
+            untrimmed_alignment, outliers
+        )
+        logging.info('Sites removed: {}'.format(sites_removed))
+        percent_removed = (
+            spruceup.get_removed_fraction(alignment_sites, sites_removed) * 100
+        )
+        logging.info(
+            'Removed {:.2f}% of sites at cutoff of {} {}s'.format(
+                percent_removed, cutoff, criterion
+            )
+        )
+        cutoff_report_fname = '{}_{}s-cutoff-{}'.format(
[…]
