+++ spruceup/spruceup.py
@@ -1,16 +1,13 @@
 #! /usr/bin/env python3
 
 # coding: utf-8
-import argparse
 import configparser
 import logging
-import os
 import random
 import json
 import pdb
 from sys import exit
 import multiprocessing as mp
-import time
 from functools import partial
 from math import log
 
@@ -19,27 +16,10 @@ import numpy as np
 import matplotlib.pyplot as plt
 import scipy.stats as scp
 from tqdm import tqdm
-import psutil
 
-from __init__ import __version__
-import aln_parsing, aln_writing
 
 plt.switch_backend('agg')
 
-parser = argparse.ArgumentParser(usage='spruceup.py <config_file_name>')
-
-parser.add_argument(
-    'config_file_name',
-    type = str,
-    help = 'Configuration file with spruceup run parameters.'
-    )
-parser.add_argument(
-    '-v',
-    '--version', 
-    action='version',
-    version='%(prog)s {version}'.format(version=__version__),
-    help='show the version number and exit'
-    )
 
 def get_tree_dist_dict(tree_fn):
     """Make a dict of all-by-all distances from input guide tree."""
@@ -709,13 +689,6 @@ def replace_seq(text, start, end, replacement=''):
     return '{}{}{}'.format(text[:start], replacement * length, text[end:])
 
 
-def print_mem():
-    """Print memory usage using psutil module."""
-    process = psutil.Process(os.getpid())
-    mem = process.memory_info().rss / 1_000_000
-    logging.info('Used {0:.2f} MB memory\n'.format(mem))
-
-
 def remove_outliers(parsed_alignment, outliers_dict):
     """Remove outlier sequences and return trimmed alignment.
 
@@ -823,99 +796,6 @@ def read_distances_dict(distances_json):
         )
         mean_taxon_distances = json.load(f)
     return mean_taxon_distances
-
-
-def analyze(
-    alignment_file_name,
-    tree_file_name,
-    input_file_format,
-    data_type,
-    window_size,
-    overlap,
-    cores,
-    method,
-    fraction,
-):
-    """Load, parse, and analyze the data."""
-    logging.info('Parsing alignment {} ...\n'.format(alignment_file_name))
-    aln_tuple = aln_parsing.parse_alignment(
-        alignment_file_name, input_file_format
-    )
-    aln_name, aln_dict = aln_tuple
-    no_missing_ambiguous_dict = replace_missing_in_dict(aln_dict, data_type)
-    if tree_file_name == None:
-        logging.info('Did not find guide tree, continuing ...\n')
-        tree_dists = None
-    else:
-        logging.info('Reading in guide tree {} ...\n'.format(tree_file_name))
-        tree_dists = get_tree_dist_dict(tree_file_name)
-    windows = get_windows(no_missing_ambiguous_dict, window_size, overlap)
-    all_distances = distances_wrapper(
-        windows, tree_dists, cores, data_type, method=method, fraction=fraction
-    )
-    taxa_distances = dist_taxa_wrapper(all_distances)
-    mean_aln_distances = mean_distances_wrapper(taxa_distances)
-    mean_taxon_distances = dists_per_taxon(mean_aln_distances)
-    write_distances_dict(mean_taxon_distances, method, window_size, overlap)
-    return (aln_tuple, mean_taxon_distances)
-
-
-def output_loop(
-    untrimmed_alignment,
-    distances,
-    window_size,
-    method,
-    criterion,
-    cutoffs,
-    manual_cutoffs,
-    report_file_name,
-    out_format,
-    out_file_name,
-    data_type,
-):
-    """Write and plot output files from analysis."""
-    alignment_sites = get_alignment_size(untrimmed_alignment)
[…]
