+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,30 @@
+Code of Conduct
+
+Having a Code of Conduct is our way to encourage good behavior and discourage bad behavior in our open source community.
+
+# Scope
+
+This Code of Conduct applies both within project spaces and in public spaces when an individual is representing the project or its community. Examples of representing a project or community include using an official project e-mail address, posting via an official social media account, or acting as an appointed representative at an online or offline event. Representation of a project may be further defined and clarified by project maintainers.
+
+# Our standards and responsibilities
+
+We encourage participation from a large group of people who bring different perspectives to the project.
+
+We expect participants to communicate professionally and thoughtfully.
+
+We call out common situations where this is especially important. For example: when people disagree, when people join a project midstream, when people are unclear in their communication style, and when you feel the need to leave the community. 
+
+We call out examples of behaviors which cause members to lose their good standing in the community. These include insulting people on the basis of their protected-class status, threatening words, or unwanted sexual attention. 
+
+
+# Enforcement
+
+Please report every incident directly to the main project developer at petiolus@gmail.com. The developer will handle your issue with discretion. We respect your confidentiality for the purpose of protecting victims of abuse.
+
+# Attribution and acknowledgments
+
+We thank those who created codes of conduct and diversity statements that served as our inspiration. This code of conduct is adapted from [Open Code of Conduct project](https://github.com/todogroup/opencodeofconduct/) and GitHub [Contributor Covenant template](https://github.com/marekborowiec/spruceup/community/code-of-conduct/new?template=contributor-covenant)
+
+This text is shared under the [CC-BY-4.0](https://creativecommons.org/licenses/by/4.0/) license. 
+
+You may consider adopting some of the following [suggested changes](./code_improvement.md) to this Code of Conduct.
