+++ codeofconduct.md
@@ -1,19 +1,0 @@
-Code of Conduct
-
-Having a Code of Conduct is our way to encourage good behavior and discourage bad behavior in our open source community.
-
-We encourage participation from a large group of people who bring different perspectives to the project.
-
-We expect participants to communicate professionally and thoughtfully. 
-
-We call out common situations where this is especially important. For example: when people disagree, when people join a project midstream, when people are unclear in their communication style, and when you feel the need to leave the community. 
-
-We call out examples of behaviors which cause members to lose their good standing in the community. These include insulting people on the basis of their protected-class status, threatening words, or unwanted sexual attention. Rather than fighting back, please report every incident directly to the project developer @marekborowiec.
-
-Please follow the reporting process so that the developer can take action on your report. The developer will handle your issue with discretion. We respect your confidentiality for the purpose of protecting victims of abuse.
-
-We thank those who created codes of conduct and diversity statements that served as our inspiration. This code of conduct is adapted from [Open Code of Conduct project](https://github.com/todogroup/opencodeofconduct/)
-
-This text is shared under the [CC-BY-4.0](https://creativecommons.org/licenses/by/4.0/) license. 
-
-You may consider adopting some of the following [suggested changes](./code_improvement.md) to this Code of Conduct.
