### 2019-10-03 f33d0ac
added contributing guidelines
