+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# Contributing
+
+When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owner of this repository (petiolus@gmail.com) before making a change.
+
+Please note we have a code of conduct, please follow it in all your interactions with the project.
