+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,34 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Method of  `spruceup` installation
+2. What environment was the bug encountered in [e.g. 'clean, new Conda environment']
+3. What command was issued
+4. Input and output files
+5. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Desktop (please complete the following information):**
+ - OS: [e.g. iOS] and OS version
+ - Python version [e.g. 3.6.5]
+ - `spruceup` version [e.g. 2019.2.2]
+- Conda and/or pip version [if applicable]
+
+**Additional context**
+Add any other context about the problem here.
