### 2019-10-03 marekborowiec/spruceup#5 @marekborowiec
Thank you very much for your careful evaluation of the source code and manuscript. Please let me know if these are satisfactorily resolved so we can move forward. Also tagging  @csoneson.

First I would like to address comments by @gavinmdouglas:

>When briefly describing spruceup at the top of the README it would be good to add a sentence stating what research gap this tool addresses. I.e. what are the advantages of this tool compared to the other tools you mention in your paper? This would help address this requirement: A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I inserted the following statement at the beginning of README:

```
`Spruceup` is a Python tool for biologists (bioinformaticians, phylogeneticists, evolutionary biologists) doing inference on phylogenomic sequence alignments. It allows discovery and removal of individual poorly aligned sequences or sequence fragments (alignment rows), which is different from the problem of poorly aligned sequence blocks (alignment columns) commonly addressed by alignment trimming software.
```

> There should be an official github release matching the pip release in the future so that it's clear what exact version on GitHub matches the source code distributed with pip.

I created a matching GitHub release (https://github.com/marekborowiec/spruceup/releases/tag/v2019.1.2) corresponding to the current PyPi release and will continue this practice.

> Installation instructions should be re-ordered so that recommended conda/pip instructions are first and "conda activate spruceup" should also be listed clearly as a separate line so that it can be copied and pasted more easily.

Done.

> I recommend that the input option / argument handling is improved. For instance, argparse could be used, which would avoid the need to use a configuration file and provide help documentation to the user. If you want to stick with the configuration file approach you need to at least point the user to where they can find the documentation for using the script if they run the command with no options or if they input "--help" or "-h" (see next point).

I believe that configuration file versus command line interface is often a matter of personal preference. I've used both in the past and I believe configuration file is more suited for this particular application. This is because it gives a more direct and immediate overview of the many options and parameters available to the user. Therefore, I will stick with configuration files.

>  At the very least the "--version" and "--help" options should be added to the script.

Versioning and help are now handled via `argparse`.

> I received the below error when installing the tool on OS X. You should provide instructions for Mac users if you intend the tool to be used on OS X.

I haven't been able to look into the OS X error yet; I believe it has to do with `matplotlib`. I added a remark about `spruceup` having been tested under Ubuntu Linux but I would like Mac users to be able to use this, so I will continue working on this.

> Add instructions for how to run unittests if installing from source

Done. I needed to restructure the repo and code for tests to work smoothly. As the result now 1) `spruceup` tests are written using `unittest` instead of `nose`, 2) `spruceup` is now intended to be used from command line as a Python module with `python -m spruceup` and 3) testing is integrated with `setuptools`. These changes are reflected in the README.

> Recommend changing criterion setting in example config file to mean if that is the recommended option.

Kept it as lognormal but removed recommendation to use mean criterion from README. Lognormal is now also discussed first in documentation for consistency.

> More details are needed for the example usage dataset. In particular, it's not clear whether this is a realistic dataset or not.

The example alignment is a fragment of an empirical dataset generated by myself and colleagues. In README file I added the following statement:
```
The example alignment is a subset of empirical, anonymized [ultraconseved element or UCE](https://www.ultraconserved.org/) data set generated from insects.
```

> The example command also failed for me with Sorry, could not read input alignment file "test_aln_nt.fasta": No such file or directory (it worked when I ran it in the examples folder)

This should not be a problem now because of more explicit instructions telling the user that this is the path rather than file name and that by default the program will look in the directory where the command was ran.

> "?" is not an IUPAC character (https://www.bioinformatics.org/sms/iupac.html) and may cause confusion for users and downstream programs. Is it intended to represent a gap or an N character? The appropriate IUPAC character should be output instead if possible otherwise a clearer explanation needs to be given when describing the output.

Many alignments used for phylogenetics use the non-IUPAC charater '?', presumably because phylogeneticists wish to differentiate between undetermined bases (coded as 'N'), gaps introduced in the process of alignment (which may correspond to deletions or insertions and coded as '-') and missing data resulting from failure to capture homologous sequences across samples ('?'). The popular phylogenetic software I'm familiar with (RAxML, IQ-TREE, MrBayes) all seems to handle '?' although to my knowledge the algorithms they employ to calculate likelihood do not differentiate between '*', '.', '-', '?', or 'N' (the last one in the case of DNA alignments). For the sake of avoiding confusion, however, ``spruceup`` now replaces trimmed sequences with the IUPAC character '-'.

Manuscript edits suggested by @gavinmdouglas:

> First sentence is very long - recommend to remove "by leveraging the large amount of data available" to be simply "from multiple sequence alignments".

Changed as suggested.

> Usage of "phylogenomic" might be misleading since this tool can work with any MSA right? I don't think this is a standard term so it might cause some confusion.

The tool can be used with any MSA but the algorithm's power to detect outliers increases with amount of data available. I am open to suggestions about the language in the manuscript but I disagree that "phylogenomic" is a rarely used term causing confusion. In the field of systematics the term "phylogenomic" is often used to distinguish large datasets from smaller ones, the latter usually assembled using Sanger sequencing. A phylogenetic dataset comprising hundreds to thousands of markers from throughout the genome would be called "phylogenomic". See for example [1](https://doi.org/10.1126/science.1157704]), [2](https://doi.org/10.1038/nature06614), [3](https://doi.org/10.1093/molbev/msw046). The search term "phylogenomic" returns >37,000 hits on Google Scholar.

> By "incorrect orthology prediction" do you mean homoplasy?

Thank you for raising this pont. In short, no. By "incorrect orthology prediction" I mean a situation where paralogous gene is recovered as a putative ortholog, potentially resulting in incorrect phylogenetic inference of relationships among species (although not necessarily sequences themselves). Since the phrase "incorrect orthology prediction" indeed appears very rarely used, I replaced it with "hidden paralogy". I'm not sure if this will be universally understood either, but at least there is a [peer-reviewed article](https://doi.org/10.1093/icb/icq044) describing the phenomenon.

> "Multiple Sequence Alignment" should be written in the same case in both the overview and summary sections.

Thank you for noticing this inconsistency; fixed.

> Add "the" before "evolutionary history of genetic material"

Done.

> Add "they are" before "derived from the same"

Done.

> Replace "Because of complexities" with "Due to the complexities"

Done.

> OTU should be defined on first usage

I dropped the "OTU" and replaced it with "sample" per your suggestion below as well as @iimog's.

> I'm not familiar with the term "subtended" - just double checking that it's the intended word

This is a valid term. It simply means "sitting on" as in "the branches subtending most genera were relatively short (...)" [(source)](https://doi.org/10.1186/1471-2148-9-56). However, I replaced it here with "characterized by" for clarity.

> Could use a sentence stating what users you envision using this tool - do you think this would be best for newcomers or experienced phylogenetics people or both?

I modified the last sentence of the Summary section to read as follows: "``spruceup`` is intended for any phylogeneticist working with large datasets as a complement to commonly used tools that filter poorly-aligned MSA sites."

> Do the sequences need to be representative sequences (i.e. OTUs)? Will the tool have trouble if there are many nearly identical sequences? If not then specifying OTUs rather than just saying "sample" or "sequence" may just cause confusion.

I have replaced all the occurrences of "OTU" with "sample" to avoid potential confusion.

> I think it's more accurate to say that it's written in Python 3, not 3.7.

Fair enough.

> A lot of the rows in Figure 1 are redundant - I think removing most of them would make this a lot easier to visualize (and you can simply mention that more sequences were actually included originally). Also more details are needed such as what "?" corresponds to.

I disagree. I believe that all rows except 2 are unique but some sites are shared across samples, just as one would expect in an alignment. Even if several rows were uniform, however, it is very common to see alignment fragments that have multiple uniform sequences even if these are not identical across the entire alignment. Because ``spruceup`` makes use of distance information (in the case of uniform sequence windows it is 0) to remove outlier sequences, I believe removing rows will be confusing because it would amount to removing information that the algorithm takes into account when identifying outliers. I do not believe that persons who work with alignments used for phylogenetic reconstruction will be confused.

As for using the non-IUPAC "?", I replaced the figure and adjusted the ``spruceup`` code, which now replaces trimmed-out sequences with '-' (see also above). I have modified and expanded the figure caption, however, to include more details. It now reads as follows:

```
Outlier removal from multiple sequence alignment by ``spruceup``. This nucleotide alignment was simulated under a phylogenetic tree with topology and branch lengths taken from the empirical dataset of Ultraconserved Elements [@faircloth_2012uces] sequenced for a group of ants [@blaimer_2016acropyga]. 100 alignments of 500 nucleotides each were simulated under the general time-reversible model (GTR) with state change rates between 0.1 and 1.0, base frequencies ranging from 0.1 to 0.8, and branch length scaling between 0.1 and 50, each parameter selected randomly for each alignment. After the 100 alignments were concatenated, divergent (complemented) sequence fragments were introduced (left panels), and later trimmed using the ``spruceup`` workflow (right panels). The figure shows only a small window of the simulated concatenated alignment of 50,000 nucleotides. Alignments visualized with AliView: https://ormbunkar.se/aliview/.
```

@iimog

> I agree that this tool is likely of high utility to the community but can be further improved
Although there are some basic guidelines for contributing, these suggestions can make it even simpler: https://github.com/marekborowiec/spruceup/community

I added a code of conduct, brief contributing guidelines, and issue templates.

> Manuscript:

> I'm a bit confused by the usage of "sample", "sequence" and "OTU" to refer to the alignment rows. Maybe it would be better to stick to one of them or clearly distinguish the different meanings.

I replaced any references to OTU with "sample". Hopefully this is now consistent and less confusing.

> I don't see how "Lack of tools for automated detection of these sequences means that few researchers attempt manual alignment curation [...]". I'd rather assume that the availability of automated tools means that few use manual curation. I agree that many will not address it at all, though.

Thank you for highlighting this potentially confusing sentence. I rephrased it to read "Lack of tools for automated detection of these sequences means that most researchers simply do not address them and only few attempt manual alignment curation".

> Duplicated it in "It it is also capable [...]"

Fixed.
