### 2019-10-03 marekborowiec/spruceup#6 @marekborowiec
This should now be addressed. Please see #5
