### 2019-10-03 marekborowiec/spruceup#4 @marekborowiec
These should now be resolved through 1. README instructions that specifically refer to example files when downloaded from source repo, 2. added explanation that the user must be running `spruceup` in the same directory as the alignment and config file or, alternatively, provide absolute paths in config and 3. default is now one core.
