### 2019-10-04 marekborowiec/spruceup#5 @gavinmdouglas
Hey @marekborowiec,

Just took a read through your response and everything looks good except I ran into an error when running the unittests (see below).

First though, here are some quick responses to specific points above:

> Many alignments used for phylogenetics use the non-IUPAC charater '?', presumably because phylogeneticists wish to differentiate between undetermined bases (coded as 'N'), gaps introduced in the process of alignment (which may correspond to deletions or insertions and coded as '-') and missing data resulting from failure to capture homologous sequences across samples ('?'). The popular phylogenetic software I'm familiar with (RAxML, IQ-TREE, MrBayes) all seems to handle '?' although to my knowledge the algorithms they employ to calculate likelihood do not differentiate between '*', '.', '-', '?', or 'N' (the last one in the case of DNA alignments). For the sake of avoiding confusion, however, spruceup now replaces trimmed sequences with the IUPAC character '-'.

Thanks for the explanation, that's good to know. I think outputting '?' would be fine as well as long as it was clearly described, but it sounds like it doesn't make a difference for downstream tools either way.

> In the field of systematics the term "phylogenomic" is often used to distinguish large datasets from smaller ones, the latter usually assembled using Sanger sequencing

Regarding the "phylogenomic" discussion - I just meant that I don't think it describes multiple-sequences alignments in general and it sounds like we agree. If spruceup works with multiple-sequence alignments of a single or a small number of genes this could be good to clarify, but that's a really minor point and not necessary to change in the pdf.

> I disagree. I believe that all rows except 2 are unique but some sites are shared across samples, just as one would expect in an alignment. Even if several rows were uniform, however, it is very common to see alignment fragments that have multiple uniform sequences even if these are not identical across the entire alignment. Because spruceup makes use of distance information (in the case of uniform sequence windows it is 0) to remove outlier sequences, I believe removing rows will be confusing because it would amount to removing information that the algorithm takes into account when identifying outliers. I do not believe that persons who work with alignments used for phylogenetic reconstruction will be confused.

Yes that's fair that removing sequences might actually cause confusion.

When running the unittests in a fresh conda environment with spruceup installed from source I ran into this error:

> ERROR: simple_tests (unittest.loader._FailedTest)
> ----------------------------------------------------------------------
> ImportError: Failed to import test module: simple_tests
> Traceback (most recent call last):
>   File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.6/unittest/loader.py", line 153, in loadTestsFromName
>     module = __import__(module_name)
>   File "/home/gavin/local/prg/spruceup/tests/simple_tests.py", line 6, in <module>
>     from nose.tools import assert_true, assert_false, assert_equal, assert_raises
> ModuleNotFoundError: No module named 'nose'

I installed `nose` with conda and then received this error when re-running the tests:

> ======================================================================
> ERROR: simple_tests (unittest.loader._FailedTest)
> ----------------------------------------------------------------------
> ImportError: Failed to import test module: simple_tests
> Traceback (most recent call last):
>   File "/home/gavin/local/miniconda3/envs/spruceup/lib/python3.6/unittest/loader.py", line 153, in loadTestsFromName
>     module = __import__(module_name)
>   File "/home/gavin/local/prg/spruceup/tests/simple_tests.py", line 8, in <module>
>     from spruceup import (
> ImportError: cannot import name 'jc_correction'

Any idea what could be going on here? Note that it works when I run `python -m unittest tests.test_simple`.
