### 2019-10-04 ef67fb2
deleted obsolete tests
