+++ tests/simple_tests.py
@@ -1,421 +1,0 @@
-#! /usr/bin/env python3
-
-from statistics import mean
-from math import log
-
-from nose.tools import assert_true, assert_false, assert_equal, assert_raises
-
-from spruceup import (
-    jc_correction,
-    p_distance,
-    get_scaled_distance,
-    get_distances,
-    get_taxon_map,
-    get_list_mean,
-    get_dist_and_taxa_lists,
-    get_dist_matrix,
-    get_taxon_map,
-    get_mean_distances,
-)
-
-# some short mock sequences
-empty_seq = '-' * 10
-partly_gappy = '---CAGTACC'
-simple_seq = 'TGTCAGTACG'
-different_seq = 'GCAATAATAA'
-short_seq = 'CAGTCG'
-simple_aa_seq = 'LNIWVNLAVC'
-different_aa_seq = 'DKVCTIVSTQ'
-
-# different dicts represent different parsed alignments
-aln_dict1 = {
-    'sp1': simple_seq,
-    'sp2': simple_seq,
-    'sp3': simple_seq,
-    'sp4': simple_seq,
-    'sp5': simple_seq,
-}
-aln_dict2 = {
-    'sp1': simple_seq,
-    'sp2': different_seq,
-    'sp3': empty_seq,
-    'sp4': empty_seq,
-    'sp5': empty_seq,
-}
-aln_dict3 = {
-    'sp1': simple_seq,
-    'sp2': different_seq,
-    'sp3': simple_seq,
-    'sp4': partly_gappy,
-    'sp5': empty_seq,
-}
-
-aln_ident = ('aln_ident', aln_dict1)
-aln_gappy = ('aln_gappy', aln_dict2)
-aln_norma = ('aln_norma', aln_dict3)
-alns_tpls = [aln_ident, aln_gappy, aln_norma]
-
-spp = ['sp1', 'sp2', 'sp3', 'sp4', 'sp5']
-zero_dist_list = [(sp1, sp2, 0) for sp1 in spp for sp2 in spp]
-
-methods = ['uncorrected', 'jc']
-
-fractions = [0, 0.5, 1]
-
-data_types = ['aa', 'nt']
-
-tree_dists = {
-    ('sp1', 'sp1'): 0.0,
-    ('sp1', 'sp2'): 0.1,
-    ('sp1', 'sp3'): 3,
-    ('sp1', 'sp4'): 0.5,
-    ('sp1', 'sp5'): 0.0,
-    ('sp2', 'sp1'): 0.1,
-    ('sp2', 'sp2'): 0.0,
-    ('sp2', 'sp3'): 0.5,
-    ('sp2', 'sp4'): 0.1,
-    ('sp2', 'sp5'): 0.1,
-    ('sp3', 'sp1'): 3,
-    ('sp3', 'sp2'): 2.5,
-    ('sp3', 'sp3'): 0.0,
-    ('sp3', 'sp4'): 0.9,
-    ('sp3', 'sp5'): 1,
-    ('sp4', 'sp1'): 0.5,
-    ('sp4', 'sp2'): 0.1,
-    ('sp4', 'sp3'): 0.9,
-    ('sp4', 'sp4'): 0.0,
-    ('sp4', 'sp5'): 0.7,
-    ('sp5', 'sp1'): 0.0,
-    ('sp5', 'sp2'): 0.1,
-    ('sp5', 'sp3'): 1,
-    ('sp5', 'sp4'): 0.7,
-    ('sp5', 'sp5'): 0.0,
-}
-
-
-def test_get_uncorrected_no_nt_distances_returns_both_aln_name_dists():
-    assert_equal(
-        len(
-            get_distances(
-                alns_tpls[0],
-                tree_dists,
-                methods[0],
[…]
