### 2019-10-04 marekborowiec/spruceup#5 @marekborowiec
Thanks a lot again for your thoughtful feedback and also for getting back on this so quickly!

I think the error happened because I failed to remove the obsolete `nose` unit tests from the remote repo... Both old and new test files were there and that's why it worked with `unittest` but not `nose`. It should be fine now.
