### 2019-10-04 52850eb
removed reference to OS and Windows system support
