### 2019-10-04 marekborowiec/spruceup#5 @gavinmdouglas
Hi again, @marekborowiec - I just checked and all the tests did pass now so that addresses all of my points. :)
