### 2019-10-04 #1635 @gavinmdouglas
All of my points have been addressed in https://github.com/marekborowiec/spruceup/issues/5 and I have nothing else to add.
