### 2019-10-05 marekborowiec/spruceup#5 @iimog
@marekborowiec well done. Thanks for addressing all the issues.
