### 2019-10-05 00d9cdb
A couple of bib fixes
