+++ paper.bib
@@ -8,7 +8,7 @@
   volume = {102},
   pages = {20--29},
   author = {Bonnie B. Blaimer and John S. LaPolla and Michael G. Branstetter and Michael W. Lloyd and Se{\'{a}}n G. Brady},
-title = {Phylogenomics, biogeography and diversification of obligate mealybug-tending ants in the genus \textit{Acropyga}},
+title = {Phylogenomics, biogeography and diversification of obligate mealybug-tending ants in the genus {\textit{Acropyga}}},
   journal = {Molecular Phylogenetics and Evolution}
 }
 
@@ -148,7 +148,7 @@ title = {Phylogenomics, biogeography and diversification of obligate mealybug-te
   number	= {1},
   pages		= {10},
   author	= {Patrick K{\"{u}}ck and Karen Meusemann and Johannes
-		  Dambach and Birthe Thormann and Bj{\"{o}}rn M von Reumont
+		  Dambach and Birthe Thormann and Bj{\"{o}}rn M {von Reumont}
 		  and Johann W Wägele and Bernhard Misof},
   title		= {Parametric and non-parametric masking of randomness in
 		  sequence alignments can be improved and leads to better
@@ -217,7 +217,7 @@ title = {Phylogenomics, biogeography and diversification of obligate mealybug-te
   pages		= {1--25},
   publisher	= {Museum National D{\textquotesingle}Histoire Naturelle},
   volume	= {283},
-  author	= {Herv{\'{e}} Philippe and Damien M. de Vienne and Vincent
+  author	= {Herv{\'{e}} Philippe and Damien M. {de Vienne} and Vincent
 		  Ranwez and B{\'{e}}atrice Roure and Denis Baurain and
 		  Fr{\'{e}}d{\'{e}}ric Delsuc},
   title		= {Pitfalls in supermatrix phylogenomics},
