### 2019-10-05 #1635 @csoneson
@iimog, @gavinmdouglas - thanks a lot for your thorough reviews!

@marekborowiec - I have sent a PR with a couple of small fixes to a few bib items. I'll get back to you shortly to finalize the acceptance.
