### 2019-10-07 marekborowiec/spruceup#5 @marekborowiec
Thank you bot, @iimog and @gavinmdouglas!
