### 2019-10-07 #1635 @csoneson
@marekborowiec - paper looks good. Here are the next steps:
- [ ] Check that the software version is consistently reported
- [ ] Create a tagged release of the current version of the software and report the version tag here
- [ ] Archive the software in Zenodo. Make sure that the title and authors of the archive are the same as those of the paper (you can also add ORCID)
- [ ] Report back the doi of the Zenodo archive here
