### 2019-10-07 #1635 @marekborowiec
The pdf looks great! One minor remark: the current release is 2019.2.2, not 2019.1.1 as indicated at the top of this page.
