### 2019-10-07 #1635 @csoneson
That's fine - we'll set the release version when the Zenodo archive is done. That will be the final version reported with the publication.
