### 2019-10-08 2dbe92f
fixed incorrect plotting of manual cutoffs
