+++ spruceup/__init__.py
@@ -1,2 +1,2 @@
-__version_info__ = ('2019','2','2')
+__version_info__ = ('2019','2','3')
 __version__ = '.'.join(__version_info__)
