### 2019-10-08 #1635 @csoneson
@openjournals/joss-eics - we're ready for you to take over!
