### 2019-10-08 #1635 @labarba
Congratulations, @marekborowiec, your JOSS paper is now published! 🚀

Big thanks to our editor: @csoneson, and the reviewers: @iimog, @gavinmdouglas — your contribution to JOSS is greatly appreciated 🙏
